Use strategic merge for cordon/uncordon
