+++ examples/Cargo.toml
@@ -158,10 +158,6 @@ name = "pod_watcher"
 path = "pod_watcher.rs"
 
 [[example]]
-name = "node_cordon"
-path = "node_cordon.rs"
-
-[[example]]
 name = "node_reflector"
 path = "node_reflector.rs"
