+++ kube-client/src/api/util.rs
@@ -26,10 +26,55 @@ impl Api<Node> {
         self.client.request::<Node>(req).await
     }
 
-    /// Cordon a Node.
+    /// Uncordon a Node.
     pub async fn uncordon(&self, name: &str) -> Result<Node> {
         let mut req = self.request.uncordon(name).map_err(Error::BuildRequest)?;
         req.extensions_mut().insert("cordon");
         self.client.request::<Node>(req).await
+    }
+}
+
+// Tests that require a cluster and the complete feature set
+// Can be run with `cargo test -p kube --lib --features=runtime,derive -- --ignored`
+#[cfg(test)]
+#[cfg(all(feature = "client"))]
+mod test {
+    use k8s_openapi::api::core::v1::Node;
+    use crate::{
+        api::{Api, DeleteParams, ListParams, PostParams},
+        Client,
+    };
+    use serde_json::json;
+
+    #[tokio::test]
+    #[ignore] // needs kubeconfig
+    async fn node_cordon_and_uncordon_works() -> Result<(), Box<dyn std::error::Error>> {
+        let client = Client::try_default().await?;
+
+        let node_name = "fakenode";
+        let fake_node = serde_json::from_value(json!({
+        "apiVersion": "v1",
+        "kind": "Node",
+        "metadata": {
+            "name": node_name,
+            },
+        }))?;
+
+        let nodes: Api<Node> = Api::all(client.clone());
+        nodes.create(&PostParams::default(), &fake_node).await?;
+
+        let schedulables = ListParams::default().fields(&format!("spec.unschedulable==false"));
+        let nodes_init = nodes.list(&schedulables).await?;
+        let num_nodes_before_cordon = nodes_init.items.len();
+
+        nodes.cordon(node_name).await?;
+        let nodes_after_cordon = nodes.list(&schedulables).await?;
+        assert_eq!(nodes_after_cordon.items.len(), num_nodes_before_cordon - 1);
+
+        nodes.uncordon(node_name).await?;
+        let nodes_after_uncordon = nodes.list(&schedulables).await?;
+        assert_eq!(nodes_after_uncordon.items.len(), num_nodes_before_cordon);
+        nodes.delete(node_name, &DeleteParams::default()).await?;
+        Ok(())
     }
 }
